requirements.txt
# Core browser automation
browser-use

# Environment variables
python-dotenv

# Web automation
selenium

# HTTP requests
requests
